Description
Task is to implement a discriminative binary classifier for detecting skin regions in images. Expected to implement the skin classifier and produce a set of experimental results to demonstrate the strengths and drawbacks of your classifier.  A generative version of a classifier is given in the following tutorial, which you can use as a starter code for your implementation:  http://cs.fit.edu/~eribeiro/teaching/tutorials/tutorial_classifiers/html/main.html (Links to an external site.)  
The code for the above tutorial can be downloaded from the following link:  http://cs.fit.edu/~eribeiro/teaching/tutorials/tutorial_classifiers/code.zip (Links to an external site.)   The discriminative classifier is described in Section 6.4.1 and Section 9.1 of Prince's book.